CSharp Random 4 Operation Coding
This time there are 3 different labels. The properties of the first two labels will be useful for numbers to be added. Another label is required for 4 different operations. Thanks to these 3 labels that produce random values, random operations are provided to the user, and they should not forget that they have a certain time while performing these operations. By The Way, Sorry for my English.